requirements.txt
mysql-connector-python==9.2.0
langchain-google-genai==2.0.9
langchain-openai==0.3.5
ydata-profiling==4.12.2
streamlit==1.42.0
tiktoken==0.8.0